pyproject.toml
[project]
name = "memgrap"
version = "0.1.0"
description = "Graphiti-powered temporal knowledge graph memory for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "graphiti-core>=0.28.2",
    "mcp>=1.9.4",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.0.0",
    "tree-sitter>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-c>=0.23.0",
    "tree-sitter-cpp>=0.23.0",
    "tree-sitter-c-sharp>=0.23.0",
    "tree-sitter-ruby>=0.23.0",
    "tree-sitter-php>=0.23.0",
    "tree-sitter-kotlin>=0.23.0",
    "tree-sitter-swift>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
markers = ["integration: requires running Neo4j instance"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
# Tree-sitter query strings have long lines
"src/indexer/language_configs.py" = ["E501"]
# Session scripts need sys.path manipulation before imports
"src/session/session_recall.py" = ["E402"]
"src/session/session_save.py" = ["E402"]
